+++ .prettierignore
@@ -1,2 +1,5 @@
 # package.json is formatted by package managers, so we ignore it here
 package.json
+coverage
+build
+types/ethers-contracts
