+++ tsconfig.json
@@ -43,22 +43,11 @@
     /* Experimental Options */
     // "experimentalDecorators": true /* Enables experimental support for ES7 decorators. */,
     // "emitDecoratorMetadata": true /* Enables experimental support for emitting type metadata for decorators. */,
-    "lib": [
-      "esnext",
-      "dom"
-    ],
-    "typeRoots": [
-      "node_modules/@types",
-      "types"
-    ],
+    "lib": ["esnext", "dom"],
+    "typeRoots": ["node_modules/@types", "types"],
     "downlevelIteration": true
   },
-  "include": [
-    "src/**/*.ts"
-  ],
-  "exclude": [
-    "node_modules/**",
-    "test"
-  ],
+  "include": ["src/**/*.ts"],
+  "exclude": ["node_modules/**", "test"],
   "compileOnSave": false
 }
