+++ tsconfig.module.json
@@ -5,7 +5,5 @@
     "outDir": "build/module",
     "module": "esnext"
   },
-  "exclude": [
-    "node_modules/**"
-  ]
+  "exclude": ["node_modules/**"]
 }
